+++ udata/harvest/tasks.py
@@ -19,15 +19,18 @@ def harvest(self, ident):
     Backend = backends.get(current_app, source.backend)
     backend = Backend(source)
     items = backend.perform_initialization()
-    if items > 0:
+    if items is None:
+        pass
+    elif items == 0:
+        backend.finalize()
+    else:
         finalize = harvest_job_finalize.s(backend.job.id)
         items = [
             harvest_job_item.s(backend.job.id, item.remote_id)
             for item in backend.job.items
         ]
         chord(items)(finalize)
-    elif items == 0:
-        backend.finalize()
+    
 
 
 @task(ignore_result=False, route='low.harvest')